# Backport DDB connector Credential Provider Fix EMR 7.1.0